Changing embedded web map JSON to legal syntax.
The "map" property of MapExtension no longer supports JSON (it's only
used to pass in an actual Map object). The webMap/webScene properties
must be used instead. Furthermore, starting with Web 5.26.0, this JSON
must fully conform to the web map spec, as it is Esri's API that will
parse it.